Move qpu and classical modules into own files
This makes the code much more navigable I think because there is so much
code for bases and vectors.
